# ProducerPerformance: reduce cost of message randomisation