Instead of generating random messages continuously when using `--record-size N` for the `ProducerPerformance` script, pre-generate approximately 10MB of candidate byte arrays of size `N`, then randomly select candidates from them as we do when we load the candidates from file.

Why:
I noticed that sending fixed data from a file with messages of size N achieved roughly 2 times the throughput of using `--record-size` N. On profiling I found that the random byte generation was using a significant amount of wallclock time. By pre-generating some randomized data we avoid the cost of randomising all the data but still get data that is not unrealistically compressible (as fixed in PR #10469)

## trunk
```
kafka-producer-perf-test.sh --topic demo_p30 --num-records 5000000 --record-size 512 --throughput -1 --print-metrics --producer-props bootstrap.servers=localhost:9092 batch.size=512000 linger.ms=100 buffer.memory=128000000 compression.type=zstd
...
5000000 records sent, 100154.237526 records/sec (48.90 MB/sec), 10.01 ms avg latency, 392.00 ms max latency, 10 ms 50th, 16 ms 95th, 19 ms 99th, 40 ms 99.9th.
...
producer-metrics:compression-rate-avg:{client-id=perf-producer-client}                      : 0.611
```
![trunk](https://user-images.githubusercontent.com/1667008/225805381-b53dcdc4-01a3-400c-93f1-c38654677dff.png)

## pre-generated random messages
```
kafka-producer-perf-test.sh --topic demo_p30 --num-records 5000000 --record-size 512 --throughput -1 --print-metrics --producer-props bootstrap.servers=localhost:9092 batch.size=512000 linger.ms=100 buffer.memory=128000000 compression.type=zstd
...
5000000 records sent, 208794.421013 records/sec (101.95 MB/sec), 6.50 ms avg latency, 326.00 ms max latency, 6 ms 50th, 10 ms 95th, 16 ms 99th, 52 ms 99.9th.
...
producer-metrics:compression-rate-avg:{client-id=perf-producer-client}                      : 0.593
```
![pre-randomised](https://user-images.githubusercontent.com/1667008/225805444-f0e40850-7e37-47e2-9219-d92964747844.png)

### Committer Checklist (excluded from commit message)
- [ ] Verify design and implementation 
- [ ] Verify test coverage and CI build status
- [ ] Verify documentation (including upgrade notes)