@robobario 2023-03-19
As an alternative there are some gains on my machine from switching to SplittableRandom, which is not threadsafe, so it avoids the cost of the AtomicLong `compareAndSet` used by `Random`. Here's some [JMH results](https://github.com/robobario/perf-rand) testing the byte array generation.

And used for the producer perf test as above with zstd and linger:

```
./core/build/distributions/kafka_2.13-3.5.0-SNAPSHOT/bin/kafka-producer-perf-test.sh --topic demo_p30 --num-records 5000000 --record-size 512 --throughput -1 --print-metrics --producer-props bootstrap.servers=localhost:9092 batch.size=512000 linger.ms=100 buffer.memory=128000000 compression.type=zstd
...
5000000 records sent, 173142.184362 records/sec (84.54 MB/sec), 7.17 ms avg latency, 351.00 ms max latency, 7 ms 50th, 11 ms 95th, 17 ms 99th, 62 ms 99.9th.
...
producer-metrics:compression-rate-avg:{client-id=perf-producer-client}                      : 0.610
```

vs trunk
```
./core/build/distributions/kafka_2.13-3.5.0-SNAPSHOT/bin/kafka-producer-perf-test.sh --topic demo_p30 --num-records 5000000 --record-size 512 --throughput -1 --print-metrics --producer-props bootstrap.servers=localhost:9092 batch.size=512000 linger.ms=100 buffer.memory=128000000 compression.type=zstd
...
5000000 records sent, 106446.393596 records/sec (51.98 MB/sec), 9.72 ms avg latency, 348.00 ms max latency, 10 ms 50th, 16 ms 95th, 19 ms 99th, 41 ms 99.9th.
...
producer-metrics:compression-rate-avg:{client-id=perf-producer-client}                      : 0.611
```