@showuon 2023-03-20
@robobario , thanks for the alternatives.
I like the [SplittableRandom](https://docs.oracle.com/en/java/javase/11/docs/api/java.base/java/util/SplittableRandom.html) way, which is the simplest change.

For the 32 chars case, I'd like to clarify something.
From the result, you said 
```
5000000 records sent, 231149.738801 records/sec (112.87 MB/sec)
```

And with our pre-generated result, which slower than the 32 chars way:
```
5000000 records sent, 208794.421013 records/sec (101.95 MB/sec)
```

I'm confused. I thought you only change the way to generate random chars. And in theory, pre-gen way should have the fastest result. Did I miss anything?