@robobario 2023-03-20
> I'm confused. I thought you only change the way to generate random chars. And in theory, pre-gen way should have the fastest result. Did I miss anything?

In the pre-generated result, it used java.util.Random to select an item from the array, so it's possibly that.