@showuon 2023-03-20
> > I'm confused. I thought you only change the way to generate random chars. And in theory, pre-gen way should have the fastest result. Did I miss anything?
> 
> In the pre-generated result, it used java.util.Random to select an item from the array, so it's possibly that.

Interesting! I'm thinking we can go ahead with [SplittableRandom](https://docs.oracle.com/en/java/javase/11/docs/api/java.base/java/util/SplittableRandom.html) way first. And then ask in community if we want to use 32 chars way to improve the producer perf tool performance, do we need a KIP (Actually, I don't think it needs one, but I'm not 100% sure if there's anyone relying on 26 chars to verify the results... etc)? Or is there any better thoughts, ... etc. WDYT?