@chia7712 2023-03-20
> Also, if it's okay for the random data to include a few more characters and we go to 32 characters (so it would be uppercase ascii letters plus some punctuation symbols), then we could make an optimisation and obtain 4 well distributed values from each int we get from the random (or 8 from a random long). When the bound is a power of 2 we can mask each byte like they do in the random implementation to get a nicely distributed value.

Nice point. two mirror comments left:

1) performance: the data are pre-generated so the cost should be much smaller than do sending (please correct me if I misunderstand anything). Hence, is it worth having more complicated code?

2) payload content: the "content" of record value will impact the compression quality (and write performance). So +1 to make good content, but the next question is "which" content is good to performance tool ... and users? It seems to me the nicely distribution is a kind of use case, and "bad" distribution is too. Hence, is it worth having such changes? Or another way of improvement is to enable perf tool to test write throughout for "different" use cases. For example: enable to configure distribution for value length and content, and thus users can define the record value according to data pattern from true use case.