@robobario 2023-03-20
> > Also, if it's okay for the random data to include a few more characters and we go to 32 characters (so it would be uppercase ascii letters plus some punctuation symbols), then we could make an optimisation and obtain 4 well distributed values from each int we get from the random (or 8 from a random long). When the bound is a power of 2 we can mask each byte like they do in the random implementation to get a nicely distributed value.
> 
> Nice point. two mirror comments left:
> 
> 1. performance: the data are pre-generated so the cost should be much smaller than do sending (please correct me if I misunderstand anything). Hence, is it worth having more complicated code?

To clarify: the two alternatives tested were:
1. SplittableRandom: trunk with `java.util.Random` replaced with `java.util.SplittableRandom`
2. SplittableRandom with 32 character optimisation: trunk with `java.util.Random` replaced with `java.util.SplittableRandom` plus this extra optimisation around the byte generation.

So these alternatives did not use any pre-generation.