@showuon 2023-03-23
> Happy for me to close this and create a PR swapping the Random implementation to `SplittableRandom` @showuon?

Sounds good. Thanks @robobario !