@robobario 2023-03-23
Closing, switching to `SplittableRandom` would give a boost without the edge cases.